Fixed it with Popen and path nonsense
