+++ src/main/test.py
@@ -1,25 +1,24 @@
-import os, glob
+import os
+import glob
 import traceback
-import asyncio
+import subprocess
 
 TEST_DIR = "./testing/"
-async def main():
-    for filename in glob.glob(os.path.join(TEST_DIR, '*.test.py')):
-        with open(filename, 'r') as f:
-            try:
-                greeting = f"\033[1m\033[92mRunning Test:\033[0m {filename}"
-                length = len(greeting) - 9
-                banner = "###" + (''.join( ['-'] * (length - 6))) + "###"
-                print("\n" + banner)
-                print(greeting)
-                print(banner + "\n")
 
-                print("", flush=True)
-                await exec(f.read())
-                print("", flush=True)
+for filename in glob.glob(os.path.join(TEST_DIR, '*.test.py')):
+    greeting = f"\033[1m\033[92mRunning Test:\033[0m {filename}"
+    length = len(greeting) - 9
+    banner = "###" + (''.join( ['-'] * (length - 6))) + "###"
+    print("\n" + banner)
+    print(greeting)
+    print(banner + "\n")
 
-            except Exception:
-                print(f"\033 Test threw exception. \033[0m")
-                traceback.print_exc()
-
-asyncio.get_event_loop().run_until_complete(main())
+    print("", flush=True)
+    try:
+        os.system(f"export PYTHONPATH=$PWD:$PYTHONPATH")
+        process = subprocess.Popen(["python3", filename])
+        process.wait()
+    except Exception:
+        print(f"\033 Test threw exception. \033[0m")
+        traceback.print_exc()
+    print("", flush=True)
